backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
websockets
ultralytics
opencv-python
numpy
reportlab
matplotlib
Pillow
python-multipart 